+++ crates/wdl-doc/theme/web-components/index.js
@@ -1,2 +1,3 @@
+import './alpine.js';
 import './sprocket-code.js';
 import './sprocket-tooltip.js';