+++ crates/wdl-doc/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 * Added doc comment support for workflows and tasks ([#656](https://github.com/stjude-rust-labs/sprocket/pull/656))
+* Added content-aware search backed by [Pagefind](https://pagefind.app/) ([#616](https://github.com/stjude-rust-labs/sprocket/pull/616)).
 
 #### Fixed